Modificato package name da DAO a dao. Aggiunta flag isInit a DBConnection.
