+++ src/it/unipv/sfw/dao/AdminDAO.java
@@ -1,28 +1,27 @@
-package it.unipv.sfw.DAO;
+package it.unipv.sfw.dao;
 
-import java.sql.Connection; 
+import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.Statement;
 import java.util.ArrayList;
 
-import it.unipv.sfw.model.utente.Cliente;
+import it.unipv.sfw.model.utente.Admin;
 import it.unipv.sfw.model.utente.Utente.Type;
 
-public class ClienteDAO implements IClienteDAO {
-	
+public class AdminDAO implements IAdminDAO {
 	private String schema;
 	private Connection conn;
 	
-	public ClienteDAO() {
+	public AdminDAO() {
 		super();
 		this.schema = "UTENTI";
 	}
 	
 	@Override
-	public ArrayList<Cliente> selectAll() {
+	public ArrayList<Admin> selectAll() {
 		
-		ArrayList<Cliente> result = new ArrayList<>();
+		ArrayList<Admin> result = new ArrayList<>();
 		
 		conn = DBConnection.startConnection(conn, schema);
 		Statement st1;
@@ -31,12 +30,12 @@ public class ClienteDAO implements IClienteDAO {
 		try
 		{
 			st1 = conn.createStatement();
-			String query = "SELECT (NOME, COGNOME, EMAIL, PASSWORD) FROM UTENTI WHERE TIPO='CLIENTE'";
+			String query = "SELECT (NOME, COGNOME, EMAIL, PASSWORD) FROM UTENTI WHERE TIPO='ADMIN'";
 			rs1 = st1.executeQuery(query);
 			
 			while(rs1.next()) {
-				Cliente c = new Cliente(rs1.getString(1), rs1.getString(2), rs1.getString(3), rs1.getString(4));
-				result.add(c);
+				Admin a = new Admin(rs1.getString(1), rs1.getString(2), rs1.getString(3), rs1.getString(4));
+				result.add(a);
 			}
 			
 		} catch (Exception e){e.printStackTrace();}
@@ -46,9 +45,9 @@ public class ClienteDAO implements IClienteDAO {
 	}
 	
 	@Override
-	public Cliente selectByEmail(String email) {
+	public Admin selectByEmail(String email) {
 		
-		Cliente result = null;
+		Admin result = null;
 		
 		conn = DBConnection.startConnection(conn, schema);
 		PreparedStatement st1;
@@ -56,13 +55,13 @@ public class ClienteDAO implements IClienteDAO {
 	
 		try
 		{
-			String query = "SELECT * FROM UTENTI WHERE EMAIL=? AND TIPO='CLIENTE'";
+			String query = "SELECT * FROM UTENTI WHERE EMAIL=? AND TIPO='ADMIN'";
 			st1 = conn.prepareStatement(query);
 			st1.setString(1, email);
 			rs1 = st1.executeQuery(query);
 			
 			while(rs1.next()) {
-				result = new Cliente(rs1.getString(1), rs1.getString(2), rs1.getString(3), rs1.getString(4));
+				result = new Admin (rs1.getString(1), rs1.getString(2), rs1.getString(3), rs1.getString(4));
 			}
 			
 		} catch (Exception e) {e.printStackTrace();}
@@ -72,7 +71,7 @@ public class ClienteDAO implements IClienteDAO {
 	}
 	
 	@Override
-	public boolean insertCliente(Cliente clienteInput) {
+	public boolean insertAdmin(Admin adminInput) {
 		
 		conn = DBConnection.startConnection(conn, schema);
 		PreparedStatement st1;
@@ -80,23 +79,22 @@ public class ClienteDAO implements IClienteDAO {
 		
 		try
 		{
-			String query = "INSERT INTO UTENTI VALUES(?,?,?," + Type.CLIENTE + ")";
+			String query = "INSERT INTO UTENTI VALUES(?,?,?," + Type.ADMIN + ")";
 			st1 = conn.prepareStatement(query);
 			
-			st1.setString(1, clienteInput.getNome());
-			st1.setString(2, clienteInput.getCognome());
-			st1.setString(3, clienteInput.getEmail());
-			st1.setString(4, clienteInput.getPassword());
+			st1.setString(1, adminInput.getNome());
+			st1.setString(2, adminInput.getCognome());
+			st1.setString(3, adminInput.getEmail());
+			st1.setString(4, adminInput.getPassword());
 			
 			st1.executeUpdate(query);
 			
 		} catch (Exception e) {
 			e.printStackTrace();
 			esito = false;
-		} 
+		}
 		
 		DBConnection.closeConnection(conn);
 		return esito;
 	}
-	
 }
